Move to class-based generic views and queryset fix on views
Previously the queryset was only fetched when the urlpatterns
were generated. This led to news items not being accessible if
they were created with a date of after this reloading took place.

Now the queryset for all date-based views is generated on the
fly in order to operate on the current timestamp.